refactor: de-dupe init_test_logger
