+++ sn_client/src/config_handler.rs
@@ -195,7 +195,7 @@ fn project_dirs() -> Result<PathBuf> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::utils::test_utils::init_test_logger;
+    use crate::utils::test_utils::init_logger;
     use bincode::serialize;
     use eyre::Result;
     use rand::{distributions::Alphanumeric, thread_rng, Rng};
@@ -211,7 +211,7 @@ mod tests {
     // 4. Verify that `Config::new()` returns the correct default config when no path is provided.
     #[tokio::test(flavor = "multi_thread")]
     async fn custom_config_path() -> Result<()> {
-        init_test_logger();
+        init_logger();
 
         let temp_dir = tempdir()?;
         let root_dir = temp_dir.path().to_path_buf();
