+++ sn_client/src/lib.rs
@@ -81,101 +81,6 @@ pub mod utils;
 /// Helpers for analysis of testnet logs
 mod testnet_grep;
 
-use tracing_core::{Event, Subscriber};
-use tracing_subscriber::{
-    fmt::{
-        format::Writer,
-        time::{FormatTime, SystemTime},
-        FmtContext, FormatEvent, FormatFields, FormattedFields,
-    },
-    registry::LookupSpan,
-};
-
-#[derive(Default, Debug)]
-/// Tracing log formatter setup for easier span viewing
-pub struct LogFormatter;
-
-impl<S, N> FormatEvent<S, N> for LogFormatter
-where
-    S: Subscriber + for<'a> LookupSpan<'a>,
-    N: for<'a> FormatFields<'a> + 'static,
-{
-    fn format_event(
-        &self,
-        ctx: &FmtContext<'_, S, N>,
-        mut writer: Writer,
-        event: &Event<'_>,
-    ) -> std::fmt::Result {
-        // Write level and target
-        let level = *event.metadata().level();
-        let target = event.metadata().file().unwrap_or("No target file known.");
-        let span_separation_string = "\t ➤ ";
-        let time = SystemTime::default();
-        write!(writer, " {} ", level)?;
-
-        time.format_time(&mut writer)?;
-
-        writeln!(
-            writer,
-            " [{}:L{}]:",
-            target,
-            event.metadata().line().unwrap_or(0),
-        )?;
-
-        write!(writer, "{}", span_separation_string)?;
-
-        // let mut span_count = 0;
-        // Write spans and fields of each span
-        ctx.visit_spans(|span| {
-            write!(writer, "{} ", span.name())?;
-
-            let ext = span.extensions();
-
-            // `FormattedFields` is a a formatted representation of the span's
-            // fields, which is stored in its extensions by the `fmt` layer's
-            // `new_span` method. The fields will have been formatted
-            // by the same field formatter that's provided to the event
-            // formatter in the `FmtContext`.
-            let fields = &ext
-                .get::<FormattedFields<N>>()
-                .expect("will never be `None`");
-
-            if !fields.is_empty() {
-                write!(writer, "{{{}}}", fields)?;
-            }
-
-            write!(writer, "\n{}", span_separation_string)?;
-
-            Ok(())
-        })?;
-
-        // Write fields on the event
-        ctx.field_format().format_fields(writer.by_ref(), event)?;
-
-        writeln!(writer)
-    }
-}
-
-use std::sync::Once;
-
-static INIT: Once = Once::new();
-
-// TODO: dedupe this w/ sn_node?
-/// Initialise logger for tests, this is run only once, even if called multiple times.
-pub fn init_test_logger() {
-    INIT.call_once(|| {
-        tracing_subscriber::fmt::fmt()
-            // NOTE: uncomment this line for pretty printed log output.
-            //.pretty()
-            .with_thread_names(true)
-            .with_ansi(false)
-            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-            .with_target(false)
-            .event_format(LogFormatter::default())
-            .try_init().unwrap_or_else(|_| println!("Error initializing logger"));
-    });
-}
-
 #[cfg(test)]
 mod tests {
     use crate::testnet_grep::search_testnet_results_per_node;
