+++ sn_interface/Cargo.toml
@@ -61,6 +61,7 @@ thiserror = "1.0.23"
 tiny-keccak = { version = "2.0.2", features = ["sha3"] }
 tracing = "~0.1.26"
 tracing-core = "~0.1.21"
+tracing-subscriber = { version = "0.3.1", features = ["env-filter", "json"] }
 uluru="3.0.0"
 url = "2.2.0"
 xor_name = "4.0.1"
